fill_keywords script to handle svn:keywords in git (SOFTWARE-2538)
In the svn sources, there are several instances where the svn:keywords
property is used on files to replace things like $Revision$ with some
text, which actually gets parsed in some places.

The application of these keywords is not consistent across the svn
sources, as (apparently) historically files were copied around with
already-filled-in keywords rather than svn-copy'd, etc.

In any case, because some places in the probes actually parse these
keyword values and break if they are not filled in, we need to fill them
in somehow for the releases based on the git sources.

Perhaps someday we'll abolish the use of these keywords altogether  :)