do a single build instead of arch+noarch (SOFTWARE-3546)
This is what we do for el7 now, and i've always been tired of needing
separate builds.

One day we'll do away with the compiled binary in the pbs-lsf probe,
and then we can make everything noarch.  But for now it's kind of fine
to call something an arch build for the one arch we support.