discard finished jobs with no elapsed time (SOFTWARE-4404)
When a partial job is reported, the DaemonStartTime is artificially
reset to the last update time, once it's gone over the cutoff pertiod.

If it so happens that the partial that we report is the last time we
see a particular job, then the remainder of the partial job will have
WallDuration = 0 (as well as total_cpu = 0, and Memory = 0).  So don't
report these.

Note that this is consistent with the previous behavior, in that exactly
what would have been reported when the job disappears already got
reported for the last partial.