cleanup 'dCache' case-hack (SOFTWARE-4621)
the issue is the directory name is mixed-case while the cron file is all
lowercase.

Renaming the probe & directory to the lowercase version is what i would
like to do, but it's unclear whether that will break things for fermi,
who is likely doing gracc queries matching ProbeName with dCache-...

So for now just use the lowercase version of the probe name here, for
the crontabs.

It feels better than the special case, anyway.