replace AuthToken with routed job attrs (SOFTWARE-5185)
Per SOFTWARE-5185, the AuthToken* attributes we used for VO info in the
htcondor-ce probe don't exist in the routed job for non-condor batch
systems.  And, even worse, these may have entirely different values for
condor batch systems (a different token can be used to authN to the local
schedd).

In HTCONDOR-1071, htcondor-ce config is added to copy the pre-routed
job's AuthToken* attrs to orig_AuthToken* attrs in the routed job.

Thus, SOFTWARE-5185 calls for replacing the AuthToken* references in the
htcondor-ce probe with orig_AuthToken*.  In order to make this
probe-specific, I add a probe_name check, as the condor-ap and
htcondor-ce probes share the same probe code.

Note also that AuthToken* references exist in common/condor_ce.py,
in certinfo code (which seems to be unimportant at this point, since
it's going away), but also in CheckAndExtendUserIdentity from
common/xml_utils.py, to fill in missing vo info for jobs by querying the
htcondor-ce directly.  Do these need to get the "orig_" prefix also?
(Always? Just when running the htcondor-ce probe?  Never?)