fix global terminationAlarm handling (SOFTWARE-5241)
I was getting an UnboundLocalError exception (local variable
'terminationAlarm' referenced before assignment), because main
sets terminationAlarm in the big try block, without calling it
global, which makes it a local variable for the whole function.

that actually means the global wasn't getting set, which is a bug,
but we happened to see it because the test later in main for
terminationAlarm references the expected-to-be-local copy in main,
which happened not to be set because we didn't make it that far
in the try block before there was an exception (apparently).

While we're at it, i'm switching the "!= None" check to the always
preferred "is not None" check.