restore /usr/share ProbeConfig symlinks (SOFTWARE-5241)
>Revert "don't bother with /usr/share ProbeConfig symlinks (SOFTWARE-4621)"
>
>This reverts commit 4326a64be445dd3b3dfa02ff33ad90014b081b42.

Apparently the /usr/share ProbeConfig symlinks were actually used by
some of the storage/transfer probes in order to refer to a default
ProbeConfig location, which was the working dir, which in the case
of these probes was the /usr/share location for the probe script
itself.  That is, the ProbeConfiguration system by default looks
for "ProbeConfig" in the current working directory, in order to
pick up the symlink to the real file under /etc.

This was not needed for the condor probes, which provide an explicit
path for the ProbeConfig; but some of the less-well-tested in 2.x
probes (basically just used by FNAL) apparently relied on the
symlink.

Now that i understand what they were for, for consistency i'm adding
the /usr/share ProbeConfig symlink to the condor probes as well, even
though they don't rely on them.