tweak ResourceName calculation (SOFTWARE-5249)
we're doing two things here:

1. add GLIDEIN_ResourceName to SYSTEM_JOB_MACHINE_ATTRS config

2. Prefer MachineAttrGLIDEIN_ResourceName0 over
   MATCH_EXP_JOBGLIDEIN_ResourceName

Per SOFTWARE-5249's description:

This will cause the AP to stuff an attribute into the job ad for each
match to a slot with the format MachineAttrGLIDEIN_ResourceNameN where
MachineAttrGLIDEIN_ResourceName0 is the slot's GLIDEIN_ResourceName for
the last job-slot match. So additionally, we need to update condor_meter
to prefer MachineAttrGLIDEIN_ResourceName0 over
MATCH_EXP_JOBGLIDEIN_ResourceName.